Intent: streamline glyph history tracking Operators involved: [Self-organization] Affected invariants: [#7] Key changes: - replace heap-based counter with Counter.most_common - simplify pop_least_used and batch trimming - refresh tests and performance checks Expected risks/dissonances: possible eviction order shifts; covered by tests Metrics: C(t), Si, νf, phase unchanged Equivalence map: None
